requirements.txt
certifi==2019.6.16
chardet==3.0.4
Click==7.0
docopt==0.6.2
docutils==0.15.2
Flask==1.1.1
Flask-SQLAlchemy==2.4.0
Hamlish-Jinja==0.3.3
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
Kivy==1.11.1
kivy-deps.glew==0.1.12
kivy-deps.sdl2==0.1.22
Kivy-Garden==0.1.4
MarkupSafe==1.1.1
pipreqs==0.4.9
psycopg2==2.8.3
Pygments==2.4.2
pypiwin32==223
python-dateutil==2.8.0
pywin32==224
requests==2.22.0
six==1.12.0
SQLAlchemy==1.3.7
urllib3==1.25.3
Werkzeug==0.15.5
yarg==0.1.9
